validate more queries, xfail some, len and window function fix

+++ python/cudf_polars/cudf_polars/streaming/benchmarks/pdsds_queries/q31.py
@@ -204,20 +204,35 @@ def polars_impl(run_config: RunConfig) -> QueryResult:
         .join(ws3, on="ca_county", suffix="_ws3")
         .with_columns(
             [
+                # TODO: DuckDB SQL promotes Decimal/Decimal to Float64; Polars keeps it as
+                # Decimal. Cast to Float64 to match DuckDB return type and filter semantics.
+                # Check other queries with Decimal arithmetic for similar semantic differences.
                 pl.when(pl.col("web_sales") > 0)
-                .then(pl.col("web_sales_ws2") / pl.col("web_sales"))
+                .then(
+                    pl.col("web_sales_ws2").cast(pl.Float64)
+                    / pl.col("web_sales").cast(pl.Float64)
+                )
                 .otherwise(None)
                 .alias("web_q1_q2_increase"),
                 pl.when(pl.col("store_sales") > 0)
-                .then(pl.col("store_sales_q2") / pl.col("store_sales"))
+                .then(
+                    pl.col("store_sales_q2").cast(pl.Float64)
+                    / pl.col("store_sales").cast(pl.Float64)
+                )
                 .otherwise(None)
                 .alias("store_q1_q2_increase"),
                 pl.when(pl.col("web_sales_ws2") > 0)
-                .then(pl.col("web_sales_ws3") / pl.col("web_sales_ws2"))
+                .then(
+                    pl.col("web_sales_ws3").cast(pl.Float64)
+                    / pl.col("web_sales_ws2").cast(pl.Float64)
+                )
                 .otherwise(None)
                 .alias("web_q2_q3_increase"),
                 pl.when(pl.col("store_sales_q2") > 0)
-                .then(pl.col("store_sales_q3") / pl.col("store_sales_q2"))
+                .then(
+                    pl.col("store_sales_q3").cast(pl.Float64)
+                    / pl.col("store_sales_q2").cast(pl.Float64)
+                )
                 .otherwise(None)
                 .alias("store_q2_q3_increase"),
             ]
@@ -225,18 +240,30 @@ def polars_impl(run_config: RunConfig) -> QueryResult:
         .filter(
             (
                 pl.when(pl.col("web_sales") > 0)
-                .then(pl.col("web_sales_ws2") / pl.col("web_sales"))
+                .then(
+                    pl.col("web_sales_ws2").cast(pl.Float64)
+                    / pl.col("web_sales").cast(pl.Float64)
+                )
                 .otherwise(None)
                 > pl.when(pl.col("store_sales") > 0)
-                .then(pl.col("store_sales_q2") / pl.col("store_sales"))
+                .then(
+                    pl.col("store_sales_q2").cast(pl.Float64)
+                    / pl.col("store_sales").cast(pl.Float64)
+                )
                 .otherwise(None)
             )
             & (
                 pl.when(pl.col("web_sales_ws2") > 0)
-                .then(pl.col("web_sales_ws3") / pl.col("web_sales_ws2"))
+                .then(
+                    pl.col("web_sales_ws3").cast(pl.Float64)
+                    / pl.col("web_sales_ws2").cast(pl.Float64)
+                )
                 .otherwise(None)
                 > pl.when(pl.col("store_sales_q2") > 0)
-                .then(pl.col("store_sales_q3") / pl.col("store_sales_q2"))
+                .then(
+                    pl.col("store_sales_q3").cast(pl.Float64)
+                    / pl.col("store_sales_q2").cast(pl.Float64)
+                )
                 .otherwise(None)
             )
         )
